s32cc: can: flexcan: Handle separate interrupt lines
On S32CC, there are separate interrupts for state change,
bus errors, MBs 0-7 and MBs 8-127 respectively.

All the IRQ numbers will be stored in the device private data instead of
the irq field of the net_device structure. That is, the interrupt number
will no longer be featured in ifconfig (as we now have multiple of
them), but we have not supported modifying it anyway.

Also, as part of this commit reform mcf5441x's irq handling to keep pace
with s32cc.
The SDK commit 8e6692f98ea4 ("can: flexcan: Handle separate interrupt
lines") remove flexcan_irq and introduces three different irq handlers.
So we need to change mcf5441x's irq handling. This is not tested for
mcf5441, only just enough to build.

Issue: ALB-3910
Issue: ALB-5979
Issue: ALB-9106
Signed-off-by: Stefan-Gabriel Mirea <stefan-gabriel.mirea@nxp.com>
Signed-off-by: Ciprian Marian Costea <ciprianmarian.costea@nxp.com>
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>