mailbox: llce: Encapsulate cached frames
The LLCE Logger uses cached frames for the cases when a new frame is
popped and doesn't belong to the channel being processed. A similar
mechanism is needed for the RX path when a FIFO is shared between
multiple BCANs. This is the preparatory work for that case.

Issue: ALB-9819
Signed-off-by: Ghennadi Procopciuc <ghennadi.procopciuc@nxp.com>