drivers/nvmem: s32cc-gpr-nvmem: Separate platform-specific offset handlers
To avoid the situation where a platform uses an NVMEM read/write
handler with an offset for a cell not corresponding to itself,
but to another platform, separated the handlers for each category
of cells (per platform), which are now called from the most
specific one to the most generic one (e.g.: s32r45 calls s32cc,
if an s32r45-specific cell offset is not matched).

The first handler to be called is determined by the compatible
against which the driver is matched when probed.

Issue: ALB-10366
Signed-off-by: Andrei Cherechesu <andrei.cherechesu@nxp.com>