thermal: qoriq: Add interrupt monitoring for one site
TMU can monitor all sites in the same way. The hardware doesn't support
defining thresholds per site. If an interrupt is activated for one
threshold, the interrupt will be fired if any of the monitored sites
breaks the limit. If only one site will break the limit, the new trips
should be updated only for that site, but this is not possible in
hardware.

This commit adds interrupts support only for one site. The rest of them
can be monitored using polling.

Taking into account that monitoring using interrupts will be enabled for
one site only, the monitoring for the other sites will be activated only
when the temperature for them is requested.

Giving the fact that the driver tries to read the temperature for a site
that was just enabled in TMSR, a delay is necessary to make sure the
updated temperature is obtained. The delay can be computed based on
TCMCFG in S32CC case, using the formula:
	1s / (tmu_freq / clk_div) * rctc * dfd * ocm

Issue: ALB-10607
Signed-off-by: Larisa Grigore <larisa.grigore@nxp.com>