gpio: siul2: configure the base
Configure the base for the GPIO driver. Even though it is deprecated to
set the base for the GPIO driver, the gpio pin ranges would not match
the pinctrl ranges if the base is not correctly configured. This means
that calls to retrieve the pinctrl driver associated to a GPIO pin might
fail.

Issue: ALB-10947

Signed-off-by: Andrei Stefanescu <andrei.stefanescu@nxp.com>