s32cc: pcie: initialize EP entries to 0
Uninitialized data allocated for EP entries in the EP list may lead to
unexpected behaviours when configuring the EP BARs, e.g. the size of an
EP bar may not be 0 before starting the configuration. Use kzalloc()
instead of kmalloc() to initialize all fields of the EP entry to 0.

Issue: ALB-11147
Signed-off-by: Bogdan Roman <bogdan-gabriel.roman@nxp.com>